Always tell #parse_expressions when to stop parsing expressions
The previous code worked fine for the examples we have so far, but it
would have allowed the end of the input to _always_ do the same job as a
close parenthesis, which isn’t what we want, strictly speaking. By
explicitly providing the pattern which terminates the list of
expressions, we can always be completely clear whether we’re expecting
the list to be ended by a parenthesis (i.e. any actual S-expression
list) or the end of the input (i.e. the top-level “list” of commands in
a .wast script).