Split out S-expression parser and tests into separate files
I’d like to start building an interpreter in wasminna.rb, and it’s going
to be easier to do that if I don’t have all the parser stuff getting in
the way. I’ve used `require` rather than `require_relative` because I’d
rather shoulder the burden of setting the load path myself than need to
go and fix up paths later on if I decide to move the files around.