Remember parameters at function definition time
I’m using a block local variable here so that I can reuse the `name`
name in order to be able to keep using hash punning. It might be better
to just extract the `map` so that I don’t have to worry about the block
polluting the outer scope of names.