Extract #with_signed helper from `div_s` and `rem_s` implementations
This helper allows us to do a computation “in the world of signed
numbers” before converting the result back to an unsigned number, which
is a pattern we use in at least two places. It looks likely that we can
also use this helper to implement `shr_s`, but it’ll need to be changed
first, because only one of the arguments to that instruction should be
interpreted as a signed number.