Use #with_signed helper in `lt_s`, `le_s`, `gt_s` and `ge_s` implementations
The helper is doing a pointless conversion of the result of #bool back
into an unsigned number, which will never actually do anything because
the #bool helper only ever returns 0 or 1, neither of which is affected
by conversion from signed to unsigned. But the conversion doesn’t do any
harm, and being able to use the helper here does make the code easier to
read and understand.