Implement the `f32.const` instruction
At the moment, the only floating-point number we know how to parse is
“nan”, which isn’t going to be enough, but it’s a start.

We’re representing 32-bit floating-point numbers as the 32-bit integer
whose bits are the same as the IEEE 754 single-precision representation
of that number. This is because:

a) WebAssembly uses IEEE 754 floating-point numbers [0]; and

b) we can already see that we’re going to need some features of IEEE 754
   floating-point numbers (e.g. NaN payloads; negative NaN) which Ruby’s
   Float class does not expose, so we can’t just use Float as our native
   representation.

This wasn’t the only option — we could’ve created some kind of data
structure for storing everything we need in addition to a Float — but
the spec talks a lot about the bitwise representation of IEEE 754
floats, so it feels as though staying close to that representation is
going to make our lives easier. But this is just a guess.

[0] https://webassembly.github.io/spec/core/exec/numerics.html#floating-point