Implement the `i32.reinterpret_f32` instruction
This is incidentally the identity function, because we’ve chosed to
represent floating-point numbers as their integer equivalent, which is
what this instruction is supposed to convert floating-point numbers
into. But if we change our minds about how floating-point numbers should
be represented internally, we’ll need to change the implementation of
this instruction to match.