Support NaN payloads
As @zetter pointed out, it’s a parsing error for the payload to be too
large to fit within the significand of the float, so it doesn’t feel
right to (silently) mask it to the correct size here. However, we’re
also not checking anywhere else that it’s an appropriate size, so in
future we might need to add error handling to the parser so that it
catches inappropriately large payloads.