Begin supporting hexadecimal float notation
So far this only works for zeros, but it’s a start. I don’t know whether
my method of turning `p`, `q` and `e` into a Float is sensible yet; it
may suffer from accuracy problems, in which case I may choose to either:

a) use BigDecimal as suggested by @zetter; or
b) concatenate the string representations of `p` and `q` before
   converting the result to an Integer, and adjust the exponent to
   compensate.

For the moment, I live in hope.