Support missing fractional part in hexadecimal float notation
I’ve made this work by coercing all matching values into strings, which
they should be anyway, but will have the effect of turning `nil` into
the empty string in the case where any of the match groups are optional,
which `q` now is. This is a bit of a hack but it works, because
`''.to_i(16)` returns zero as desired.