Use BigDecimal instead of Rational to represent float significand and exponent
The previous approach worked for almost all the tests, but failed on
`f64_dec.max_subnormal`. By using a BigDecimal instead of a Rational
when assembling the float, we avoid the (mysterious) loss of precision
that was causing this value to get rounded up to the smallest normal
number.

Again I don’t really understand what’s going on here, but I’m choosing
to prioritise making progress over fully understanding the problem.