Skip `module binary` and tolerate the resulting failure to find a function
We can’t parse these binary modules at the moment, so the quickest
solution is to just ignore them and not allow the missing functions to
cause the test to fail. Of course other mistakes might cause functions
to be missing, but hopefully the warning is enough to alert the user to
what’s happening.