Remove `bits` guards to support `i64.trunc_f*_*` instructions
The width of the resulting integer doesn’t matter in the cases we’re
currently trying to make work, i.e. when the output is defined. When we
start supporting `assert_trap` we’ll need to use the integer width to
check that the truncated float is within the range of the resulting
integer, but we’re not doing that yet.