Support subnormals in Wasminna::Float.decode
For zeros the exponent really doesn’t matter, but we’re leaving it as
literally zero because that’s a nice canonical default. For subnormal
numbers, though, we need to adjust the exponent to account for both the
bias and the encoding format, i.e. the entire significand is a fraction.