Extract a Wasminna::Float::Finite class and move #encode onto it
We need to support decoding infinities, and we can’t do that while we’re
representing floats as a triple of `numerator`, `denominator` and
`negated`. Having instances of user-defined classes will let us store
whatever we need to inside the instance, and we just have to arrange
for #encode to do the right thing in each case. Polymorphism!