Limit float exponent during significand scaling
Previously the loop was allowing exponents to become arbitrarily large
or small, which won’t work, so we need to stop the loop if that’s about
to happen.

I think the min_exponent is probably too low for subnormals, but it’s
made some of the tests for zeros pass, so I’m committing it like this
for now.