Extract Wasminna::Float::Format class
This class represents the knowledge we have about the format of 16- and
32-bit IEEE 754 floating-point numbers, i.e. the sizes of their
exponents and signifiands. This was previously repeated in several
methods, and is now consolidated in one place.

I extracted `exponent_bias` and `fraction_bits` at the same time,
because they lived directly next to the code that determined the
exponent and significand widths, but they’re arguably less important
since the caller could compute them from the other information if it
wanted to.