Return correct range of exponents from Format#exponents
Because this was extracted from Finite#encode, it actually didn’t
include the very minimum and very maximum exponents (because those have
special meanings in IEEE 754), which is both confusing and actively
inconvenient for other people who want to use the actual minimum and
maxmimum exponents precisely because they are special. It’s therefore
better to return the accurate range here, and correct for that in the
one place (i.e. Finite#encode) that needs something different.