Make Format#pack and #unpack deal with unbiased exponents
This way — at least in principle — no other code needs to know about
exponent biases. However, there’s enough code in #decode which depends
upon the biased exponent that I’m not going to try to expunge all
knowledge as part of this commit. It’ll be in an upcoming commit though.