Don’t care about exponent when decoding zero
I’d gone to the trouble of making sure the denominator was something
nice when we decoded a zero, but it really doesn’t matter what it is, as
long as the numerator is zero. We can make this code much simpler by
just allowing the “natural” exponent to carry through to the
denominator.