Move #encode out of Wasminna::Float.parse
We don’t need to know the format at all to parse the float, so we
shouldn’t take it as an argument. Before Wasminna::Float existed we had
no choice but to return the float encoded as an integer because we had
no other way of representing it, but now we can return a Wasminna::Float
instance instead.