Use Wasminna::Float everywhere in interpreter
This removes the last occurrences of Array#pack and String#unpack from
the interpreter, which is a big improvement in readability. At the
moment we’ve just moved the packing & unpacking into #to_f on the Float
classes, because I have not yet thought of another way to round-trip
conversions between Ruby and Wasminna floats.