Make float regular expressions a bit more compact
I think this strikes a better balance between clarity and vertical
length. I’m not sure about the whitespace between the capture group
names and bodies (the breathing space is good, but it prevents the
reader from being able to visually split the regular expression on
whitespace to identify its component parts), so I might revisit this in
future.