Use descriptive names instead of `p`, `q` and `e`
The original names came from the WebAssembly spec [0], but it’s clearer
to use more descriptive ones. I’m not sure that `fractional` (vs
`fraction`) is the right choice, but it’s easy enough to change my mind
later if I develop a stronger opinion.

I also changed the order of these names in the hash pattern. I think I’d
previously left `p`, `q` and `e` at the end because they were the ones
from the spec, but it’s a little less surprising for them to appear in
the same order as they do in the regular expression itself.

[0] https://webassembly.github.io/spec/core/text/values.html#floating-point