Implement the `f32.round` instruction
Thank you again @paracycle for noticing that the spec wants `f32.round`
to round to even, but Ruby rounds to zero by default.

This instruction completes all of the requirements for f32.wast, so I’ve
moved it from the pending tests to the passing ones.

Thank you also to @tekin for pointing out that, while it is useful to
remove f32.wast from the list of pending tests, it’s even more useful to
add it to the list of passing tests.

I am quite sleepy.