Combine `floor`/`ceil`/`trunc`/`nearest` implementations
These are almost identical in their handling of special values,
differing only in the underlying operation. We don’t actually need any
special value handling for `floor` (because you can’t get a `floor` of
zero when you start with a negative input), but it doesn’t do any harm,
so it’s worth including for the sake of regularity.