Combine `min`/`max` implementations
These are essentially identical in their handling of special inputs,
differing only in whether they call Array#min or Array#max (or
Array#min_by/#max_by in the case of zeroes). At the moment I’m being a
bit lazy and using #send to pick the right method, since their names
exactly match the WebAssembly instructions, but in future it might be
better to maintain a lookup table to make the code even more explicit.