Return an Enumerator from SExpressionParser#parse
This means that the interpreter can start receiving expressions straight
away by calling #each on the result of #parse, instead of waiting for a
complete array of all expressions. So tests will start running sooner,
which feels more responsive, even though the total amount of work (and
thus total runtime) is the same.