Don’t round quotients whose remainder is zero
This was the cause of a nasty bug: when we encoded a float like
0x1.7e2c44058a799p+52, whose magnitude is in the right range to not get
scaled up or down during encoding, the rounding procedure was adding one
to the quotient because the number is odd, and round-to-even is
triggered by the remainder (in this case zero) being equal to half of
the denominator (in this case also zero, because 1 / 2 = 0).

Performing any round at all is clearly wrong in the case of an integer,
i.e. a rational with a denominator of 1.

The fix is to avoid rounding integers, and “the remainder is 0” feels
like the most appropriate way of characterising that situation to me
(versus e.g. “the denominator is 1”, which would also fix the bug but
feels more disconnected from the source of the problem).