Use `2 ** exponent` to scale a Rational, per @zetter’s idea
This is a much more direct translation of how to assemble a number from
its floating-point components; we couldn’t do this originally because we
didn’t have instances of Rational and Sign to hand, but now that we do,
it’s much clearer than messing around with conditional division. Thank
you Chris for suggesting this!