Create empty function and mutate it to fill it in during definition
This more closely matches the way the language works: each function
definition is a sequence of (mostly optional, in some cases multiple)
different kinds of information about that function, and it’s hard to
write all possible pattern matches to support them all. It’d be much
better to just “interpret” the function definition, one expression at a
time, which I will begin doing in the next commit, but first we need an
empty function definition to mutate.