Implement `invoke` command outside of any assertion
For the first time, we’re running a test script (i.e. float_memory.wast)
that wants to be able to invoke a function for its side-effects, not its
return value, so we need to be able to support that.

I’ve copied and pasted a big chunk of code here, so a method needs
extracting fairly soon. I just want to make the test pass first.