Implement the `i32.store` instruction
I’m forcing the string encoding to be `ASCII_8BIT` so that we can refer
to the correct byte offset within the string. If it was `UTF_8` then
there’s a risk that there might be a multibyte character in the string,
in which case our offset (which is in bytes) would be incorrect.

Although String#byteslice allows you to *read* a substring at a given
byte offset and length, I don’t know of any similar operation which
allows you to *replace* a substring in the same way. So I’ve gone with
this hack instead.