Extract Memory#load and #store
I’m not sure whether `value` should be a keyword argument or not,
because I don’t really have a system for this. While I was working on
this I noticed that #mask has a positional argument (`value`) and a
single keyword argument (`bits:`), so something’s inconsistent here, but
at the moment I’m not sure which I prefer, so I’m just going to leave it
for now.