Reimplement Memory#store with String#setbyte
This allows us to address the underlying string as a raw array of bytes
without having to worry about its encoding, which I’m much more
comfortable about than using character indexes and messing about with
encoding. It’s also nice to be spared the details of the cryptic #pack
format strings, and express the little-endianness of the memory more
directly.