Be explicit about Integer#chr encoding in #parse_string
This is the default anyway (I think? the documentation doesn’t say, but
practically it does appear to be) but it makes the code a little clearer
to be explicit about the fact that we’re forcing the encoding of the
string so that the substituted character is compatible with it.