Rename Memory.for to Memory.from_string
So that I can add another factory for creating a blank memory with
minimum and maximum size.

It’s possible that we should only have one factory and then implement
some instance methods for things like copying a string into the memory,
but I’ll come back and clean this API up once the tests are passing.