Implement `(memory 1 1)` and similar
I can’t find the exact part of the spec that confirms this, but the
syntax of limits [0] suggests very strongly that the minimum and maximum
sizes are unsigned 32-bit integers.

At the moment I’m just ignoring the maximum size, on the assumption that
I’ll need to pay attention to it for the `memory.grow` [1] tests.

[0] https://webassembly.github.io/spec/core/text/types.html#limits
[1] https://webassembly.github.io/spec/core/exec/instructions.html#xref-syntax-instructions-syntax-instr-memory-mathsf-memory-grow