Add locals to function definitions
These aren’t yet used when executing the function — they’ll need to be
combined with the parameters to get the total set of locals for the
function body. But I’m going to defer thinking about that until I have a
failing test that requires it, otherwise I’ll have no idea whether I’ve
implemented it correctly.