Implement the `local.set` instruction
I now realise that storing the locals in a hash is not going to work in
general, because it’s possible for some (or all) of them to be
anonymous, e.g. `(local f64 f64 f64 f64)`. But the tests are currently
failing on some code where all the locals have names, so I’m not going
to fix that anticipated problem until it becomes a real one.