Support numeric branch label in `br_if` and `block`
At the moment this will only work for the label `0`, which jumps out of
the immediately enclosing block. That’s enough to get the current test
to pass, but we’ll need a more general solution for code that uses
numbers other than `0` (i.e. jumping out of an outer block or loop).