Support missing expected value in `assert_return`
The documentation for the reference interpreter [0] says that
`assert_return` can provide an arbitrary number of expected values, or
none at all. It looks like expressions are able to return multiple
values, which should be matched pairwise with the multiple expected
values, but we haven’t implemented anything that returns multiple values
yet so it’s premature to try to get that working.

Instead I’ve relaxed the pattern match to allow zero expected values,
because we need that for an `assert_return` in float_exprs.wast which
(strangely) invokes a function solely for its side effects.

[0] https://github.com/WebAssembly/spec/tree/main/interpreter#scripts