Push `const` values onto the stack
For now, we keep returning the result for compatibility with those
instructions (i.e. all of them) which currently expect the result to be
on the Ruby call stack.

Note that we don’t want to push *all* evaluation results onto the stack
indiscriminately, because some instructions don’t return anything.