Retrieve `add` operands from stack
Somehow this works for all uses of `add`, although I don’t understand
how that works in int_exprs.wast, where one of the arguments is a
`local.get` rather than an integer operation or a `const`. I’m pretty
sure we actually need `local.get` to be pushing its result onto the
stack for everything to be truly working correctly, so I’ll do that
next.