Change #invoke_function to evaluate the entire body at once after unfolding
I think this is the correct way to think about evaluating the body of a
function which contains many — possibly folded — expressions.

The bad news is that the #evaluate method currently can’t cope with
evaluating a list of instructions; it can only evaluate a single
instruction at a time. The good news is that the only tests that have
multiple instructions in their body are pending tests, so we haven’t
actually broken anything important by making this change.