Support unfolded instructions in #evaluate
When an instruction has been unfolded, the result is multiple
instructions being sent to #evaluate, which it didn’t know how to
handle. By checking for an initial folded instruction in that list, we
can at least support unfolding instructions which leave their arguments
folded. This won’t be true in general, but I’m hoping that during the
transition to evaluating unfolded expressions I can leave brackets
around the unfolded expressions to make them compatible with #evaluate.