Consume the correct number of arguments for plain instructions in #evaluate
I’m not yet able to deal with `block`, `loop` and `if` (because I
haven’t thought about it) so I’m leaving those alone in this commit.
That means that the final `rest` array does not necessarily contain only
the unused parts of the expression, because `block`, `loop` and `if`
currently don’t reassign it after using elements from it. That’s not a
problem in practice because nobody is actually using `rest` yet, but I
need to fix it before I can start using `rest`, i.e. in a loop that runs
until `rest` is empty.