Define #consume_structured_instruction method
We need this method to recursively consume an entire `block`, `loop` or
`if` instruction in the #evaluate method, so that we can correctly
attribute each `end` atom to its matching `block`/`loop`/`if` even when
those constructs are nested.

I’m concerned that this method will not be enough to consume an `if`
instruction, because those instructions have the same nesting problem
with the `else` atom, but for now we apparently don’t have any `if`
instructions with multiple instructions inside their consequent or
alternative, so it’s not a problem yet. (It presumably will be once we
flatten the unfolded expressions.)