Define #split_on_else method
As explained in e8ceb1a842183df224ff99ea7514319464402355, the existing
method #consume_structured_instruction doesn’t deal with the `else`
atom inside the “body” of an `if` instruction. To find the correct
`else` in the presence of either nested `if`s or `if`s with multiple
instructions in their consequent or alternative, we need to recursively
consume the entire “body” of the `if` and find the matching `else`.
That’s what this method does, and I’ll hook it up to the evaluation of
`if` instructions in the next commit.