Deal with nilable array elements more neatly in #unfold
I’m trying here to introduce a standard way of dealing with
potentially-nil instructions and sequences of instructions: #compact for
the former, and safe navigation for the latter. This avoids coming up
with ad-hoc solutions (like `label_and_type`) for every situation, which
I’m hoping will make the code easier to understand in aggregate.