Always return a sequence of instructions from #unfold
The `else` case is here so that we can deal with non-folded instructions
appearing anywhere that a folded instruction might: we just allow it to
pass through unchanged. But because #unfold can in general return
multiple instructions as a result of the unfolding process, it’s more
consistent to make sure that we wrap the single instruction in an array
before returning it, so that the caller doesn’t need to deal with a
mixture of individual instructions and arrays of instructions. (This
isn’t a big deal in practice because Ruby’s operations, e.g.
Array#flat_map, are pretty forgiving on this point. So it’s really the
principle of the thing.)