Consume `consequent` and `alternative` directly for `if`
We now have a flexible enough implementation of
the #consume_structured_instruction method that we can reuse it here
instead of having a whole separate method for splitting the body on
`else`. It’s going to be easier to understand and maintain this code if
we’re not duplicating the same logic three times with minor changes.