Parse `local.get` instruction
Eventually we’ll want to look up the index in the `locals` space of the
identifier context [0], but I don’t have an identifier context yet
because I’m only parsing expressions at the moment, not module
definitions (which contain the function definitions that create local
variables). For now I’m going to allow both string and integer indexes,
and let the interpreter deal with it; in future I’ll come back and clean
this up so that the index is always an integer.

[0] https://webassembly.github.io/spec/core/text/modules.html#indices