Parse `block`, `loop` and `if` instructions
This is a little more complicated and requires moving
the #consume_structured_instruction helper over from the interpreter,
but it’s still not too bad.

We’re defaulting the label to `0` as a hack to support cases (e.g. the
`tau` function in float_exprs.wast) where the label is missing but the
branch inside the block specifies index 0. The label isn’t even supposed
to survive the parsing process — the labels just populate the identifier
context, and any references to them turn into numbers when looked up in
that context — so this whole lamentable arrangement is just a
transitional bodge until we get identifier contexts working.