Say “opcode”, not “instruction”, when referring to atoms containing instruction names
They’re not instructions at all — they’re raw strings from the
S-expression, and instructions are what we emit from the AST parser.

@zetter notes that the WebAssembly spec uses the word “opcode” to refer
to the byte which identifies each instruction [0] in the binary format.
So it’s not quite the right word to use here — at best it’s overloading
a word which already has another, official meaning — but it’s better
than the actively misleading “instruction” so I’m going to go with it
for now. Hopefully I’ll think of an even better word later.

[0] https://webassembly.github.io/spec/core/binary/instructions.html