Store the current S-expression as state instead of always passing an argument
This is slightly complicated by the fact that #parse_expression
recursively calls itself (indirectly), so we need to save the previous
value of the `s_expression` attribute and restore it after the
recursive call. This isn’t currently a very tidy solution because
the #parse_expression method still takes an `s_expression` argument,
unlike all of the other #parse_* methods. I’ll try to fix this next.