Extract #parse_nested_expression
We’re doing the same pattern in several places, so it makes sense to
wrap it up in a helper. Ultimately we could use this as our primary
means of parsing expressions, but only if #read_structured could cope
with hitting the end of the input instead of finding a terminator.