Use a special object to mark the end of the whole S-expression
This allows us to avoid the conditional on `nil`, because we now always
provide a `terminated_by` (cf SExpressionParser#parse_expressions)
instead of allowing `nil` to have a special meaning.

This has performance implications because we now do a full scan through
the entire S-expression before we even start parsing anything, but I
think the clarity benefits outweigh the performance costs. If this later
turns out to be a legitimate performance bottleneck then I can mitigate
it by moving to Enumerator instead of arrays of atoms.
(SExpressionParser#parse already produces an Enumerator anyway, so we
could just consume that here.)