Handle multiple anonymous function parameters correctly
Inside a function definition, a parameter declaration can either be a
single named parameter [0] or multiple anonymous parameters [1], so we
need to support both explicitly.

This is further evidence that we need to prioritise parsing the module
S-expression in the AST parser so that the interpreter doesn’t need to
deal with details like this.

[0] https://webassembly.github.io/spec/core/text/types.html#text-functype
[1] https://webassembly.github.io/spec/core/text/types.html#abbreviations