Hack Return to empty the stack before returning a value
This makes at least one test pass (`as-br_if-last` in if.wast), but it’s
not the right behaviour in general.

The spec says [0]:

> 1. Let F be the current frame.
> 2. Let n be the arity of F.
> …
> 4. Pop [n results] from the stack.
> …
> 6. While the top of the stack is not a frame, pop the top element from
>    the stack.

This means that we shouldn’t _necessarily_ empty the stack, only remove
everything that wasn’t already there when the function was invoked. But
at the moment we don’t have enough context to be able to do that, so I’m
just hacking it to make the stack empty, which is enough to make the
current test pass. We’ll obviously need to improve this when we try to
get e.g. return.wast passing, but for now I’m trying to focus on
the minimum changes necessary to get if.wast passing.

[0] https://webassembly.github.io/spec/core/exec/instructions.html#xref-syntax-instructions-syntax-instr-control-mathsf-return