Unfold `br_table` instruction
The `br_table` instruction is unusual in that it doesn’t have a fixed
static arity: it’s followed by zero or more target labels and a
mandatory default label [0], i.e. one or more labels. We therefore need
to keep reading atoms until we find one that is neither a numeric index
nor an id.

[0] https://webassembly.github.io/spec/core/text/instructions.html#control-instructions