Add result types to Function
We need to know how many values the function returns in order to
correctly implement the `return` instruction, so we need to remember the
return types during function definition.

We could just increment a counter, but I assume these strings are going
to be useful once we start thinking about validation (i.e. type
checking). But note that I don’t care enough to have turned the strings
into `Type` objects or something — that’s too speculative at the moment.