When branching, throw a special value (`:branch`) instead of the label
We need to do this to support branches that target anything other than
their enclosing block: if the branch throws a 2, for example, we still
need to catch it in each ancestor block in turn and decrement & rethrow
it so that a 0 eventually hits the correct ancestor.

This doesn’t break any of the existing tests, so apparently we’re always
branching to the nearest enclosing `block`/`loop` in all of the
currently-passing tests, i.e. nobody cares about the label.