Use result of `catch` block to detect whether `throw` has happened
Previously we were using the value that got thrown by `throw`, but I’d
like to use that for something more important (i.e. the index, which is
either zero or greater than zero). Fortunately we can achieve the same
effect here without caring about what value got thrown, just by
inverting the sense of the check and looking for a special value that is
returned if the `throw` _doesn’t_ happen.