Rethrow branch tag in `block` and `loop` if this instruction is not the target
If the label doesn’t match, or if the index isn’t zero, the branch was
targeted at a parent scope rather than us, so we should rethrow it and
allow our immediate to handle it (and potentially rethrow again if
necessary). By decrementing the index we ensure that it will eventually
hit zero once it reaches its target scope.