Unwind the stack correctly after a branch
As explained in the previous commit, we need to unwind the stack to the
correct height after a branch (i.e. to what it was when we encountered
the label) and restore the right number of operands depending upon the
block’s arity.

Although this is really part of the operational semantics of the branch
instructions (e.g. `br`, `br_if`), it’s easier for us to implement it as
part of the instruction targeted by that branch (e.g. `block`, `loop`)
because the latter already has all of the necessary information needed
to rewind & reconstruct the correct stack, whereas the branch
instruction itself does not know which structured instruction is being
targeted.

This change specifically causes the `as-br_if-value-cond` test from
br.wast to start passing, although this exposes new failures relating to
the fact that we aren’t treating `if` as a block, which we should be.
I’ll fix that next.

I’m conscious that I’m introducing lots of duplicated code here, but
it’s becoming increasingly obvious that there’s a single “enter a
structured instruction” concept that will be shared between `block`,
`loop` and `if`, so I’ll continue hacking away for now and extract that
idea into a helper once I’m convinced everything is working correctly.