Treat If as a structured instruction for branching purposes
Previously we weren’t even trying to #catch the evaluation of the
consequent and alternative of a conditional, so any branch was
inadvertently targeting the wrong (possibly nonexistent) ancestor block.
Now we’re handling thrown labels in the same way as for Block and Loop,
so branches now target the correct ancestor block and we’re also, as a
bonus, doing the stack unwinding correctly too.

This does make the `as-if-then` test pass in br.wast, but it also
introduces yet more duplication, so it’s time for some refactoring.