Remember stack height each time we enter a loop
This isn’t necessary, but nor is it a problem, because the stack height
will be the same at the start of every iteration of the loop as a
consequence of the way loops work (i.e. you have to branch to return to
the start of the loop, and branching unwinds the stack).

I want to make this change so that it’s possible to extract the body of
this loop and share it with the duplicated code in Block and If.