Extract #as_branch_target helper
All three structured instructions (`block`, `loop` and `if`) have to
work as the target of a branch instruction (e.g. `br`), so we need the
`catch` handling and stack unwinding logic for all three of them. This
helper encapsulates that logic in a single place, and allows it to be
shared between the implementations of the structured instructions.

The #as_branch_target helper returns a boolean to indicate whether a
branch occurred or not. We need this so that the implementation of
`loop` knows whether to terminate the loop or not. I’d previously done
this with a `break`, but that’s not going to work across methods (or
even within a nested block) so we need to do it with data instead. I’d
like to find a neater solution to this — maybe some kind of callback?
I’m not sure how that would work either. Cool story.