Unfold the `call_indirect` instruction
This is slightly fiddly because of an abbreviation [0] which allows the
table index to be omitted. I’ve decided (for now) to not expand any more
abbreviations inside #unfold, because doing so for `if` has created a
situation where the abbreviation is supported for folded instructions
but not for already-unfolded ones. So I’m not inserting the default
table index (zero) into the unfolded expression; instead, I’m allowing
the actual abstract syntax parser to deal with the abbreviation in the
final unfolded expression.

[0] https://webassembly.github.io/spec/core/text/instructions.html#abbreviations