Parse the `i64.load8_s`, `i32.store8` and `i64.store16` instructions
At the moment we’re throwing away the storage size and sign extension
mode [0] because the test we’re working on, br.wast, doesn’t even try to
evaluate these instructions, so it’s good enough to be able to
successfully parse them into a Load or Store instruction.

In future we’ll need to actually remember the details of how many bits
to read or write, and whether to interpret reads as signed or unsigned
numbers for the purposes of padding them out to a larger number of bits.
We’ll cross that bridge when we come to it, but I imagine I’ll just
store the `operation:` on the AST node in the same way as I do for e.g.
`UnaryOp` and `BinaryOp`. At some point we should parse all of these
meaningful instruction names in the AST parser, but that should be a
change that affects all such instructions, not done piecemeal.

[0] https://webassembly.github.io/spec/core/syntax/instructions.html#memory-instructions