Remove empty stack assertion from #invoke_function
This was correct when we could only invoke functions from the top level
(i.e. an `invoke` action in the test script), but now that the `call`
instruction also uses #invoke_function, we need to support situations
where evaluation is halfway through computing some intermediate result
and therefore does have extraneous operands on the stack. It is no
longer true in general that the stack should be empty after the
function’s arguments have been popped.