Execute function body as branch target
This allows normal branch instructions (e.g. `br`) to target the
function body rather than an explicit structured instruction (i.e.
`block`/`loop`/`if`).

This is essentially all that `return` does, so in principle we can unify
the two mechanisms (and therefore share all the logic about stack
unwinding, which is currently not implemented correctly for `return`
even though it works for `br` et al), but at the moment I can’t see an
elegant way of doing that. I’ll come back to it.