Store function identifier and exported name separately
These are effectively two separate namespaces: functions have identifier
inside a module (used by `call`), but they can also be exported from
that module with names (used by `invoke` and `assert_return`). Some
functions, e.g. `fac` in call.wast, can have both, so we need to store
both to avoid one clobbering the other.

Ultimately these functions should just be stored in a hash keyed on name
rather than knowing their own names, but this change is simpler and gets
the job done for now, which appears to be the spirit of this particular
project.