Make loop more like block
Previously, the Wasm loop instruction was implemented with a Ruby 'loop' keyword.
Loops in Wasm are different to loops in Ruby - they exit by default and only repeat if branch to them within the loop. If you're used to loops in Ruby (and similar languages), this is rather counter-intuitive. To make this work with the 'loop' keyword the code had to break out of the loop if no branches occurred. This makes my head spin and even led to a head-in-the-hand moment when it defeated @tomstuart's attempt at a refactor[1]

Instead, lets try to handle loop instructions like blocks! The difference between blocks and loops in Wasm is that:
+ branching to a block braches to the *end* of the block
+ branching to a loop branches to the *start* of the loop
The Wasm spec calls this a 'forward jump' and 'backward jump' respectively[2]

The good news is that Ruby has a way to 'backward jump' - the 'redo' statement. The slightly bad and awkward news is that it only works within blocks, so I've used `define_method` to make the `as_branch_target` method a block and redo-able. Maybe there's a nicer to do this. Also note a similar thing could be achieved with the 'retry' keyword if we threw and rescued an exception.

I like how this removes the need for the Ruby loop and makes the code better reflect the key difference between Wasm blocks and loops. As a bonus, the return value of `as_branch_target` is now not needed anywhere.

[1] https://youtu.be/gUuRVQyehL4?t=9442
[2] https://webassembly.github.io/spec/core/syntax/instructions.html#control-instructions