Unfold instructions during instruction parsing, not up-front
This is the point at which we know for sure that we’re about to try to
parse a folded instruction, so we’ll only try to do unfolding when it’s
actually appropriate.

The problem I’m trying to solve is that flat instruction sequences like
`call_indirect (type $proc)` get messed up by unfolding because the
`(type $proc)` part looks like a folded instruction but isn’t. So by
only trying to unfold when we know we’re in “instruction position”,
instead of recursively unfolding everything indiscriminately, we should
be able to avoid touching immediate arguments like `(type $proc)`.

To do that correctly we need more control over when and where unfolding
happens, which is what this commit gives us. We still need to change
the #unfold method to leave subexpressions alone — if any of them are in
fact folded instructions, they’ll get their turn to be unfolded when
they’re parsed.

I’m unhappy about the `s_expression.unshift` situation because I don’t
think we should be pushing anything back onto the input. I’ll think
about an alternative shortly.